saque validate function del componente profile
